// attendra_fixed_V5/src/types.ts
export type UserRole = 'super_admin' | 'admin' | 'employee';
// ...
export type AttendanceStatus =
  | 'present'
  | 'late'
  | 'grace_period'
  | 'wfh'
  | 'on_leave'
  | 'early_exit'
  | 'absent'
  | 'half_day'
  | 'official_tour';

export type LeaveType = string;

export interface LeavePolicy {
  id: string;
  name: string;
  yearlyQuota: number;
  colorTag: string;
}

export type LeaveStatus = 'approved' | 'pending' | 'rejected';

export type EmploymentType = 'full_time' | 'contract' | 'remote' | 'hybrid';

export interface Employee {
  id: string;
  name: string;
  role: string;
  designation?: string;
  department: string;
  email: string;
  phone: string;
  avatar: string;
  status: 'active' | 'on_leave' | 'inactive';
  employmentType: EmploymentType;
  location: string;
  joinDate: string;
  shift: string;
  leaveUsed?: Record<string, number>;
  leaveBalance: {
    annual: number;
    sick: number;
    casual: number;
    emergency: number;
    unpaid: number;
    maternity: number;
    paternity: number;
    half_day: number;
    [key: string]: number;
  };
  manager: string;
  dob?: string;
  nidNumber?: string;
  address?: string;
  bloodGroup?: string;
}

export interface AttendanceRecord {
  id: string;
  employeeId: string;
  employeeName: string;
  employeeAvatar: string;
  department: string;
  date: string;
  entryTime: string;
  exitTime: string | null;
  workHours: number;
  overtimeHours: number;
  lateMinutes?: number;
  earlyExitMinutes?: number;
  status: AttendanceStatus;
  locationType: 'Office HQ' | 'Remote - Home' | 'Client Site' | 'Geofence App' | 'Official Tour';
  verificationMethod: 'FaceID Biometric' | 'Mobile GPS Geofence' | 'Web Wi-Fi Checkin' | 'RFID Badge' | 'Manual Admin Entry';
  reason?: string;
  notes?: string;
}